LPS-175547 Moves necessary methods to AssetEntryItemSelectorViewDescriptor
