+++ modules/apps/asset/asset-browser-web/src/main/java/com/liferay/asset/browser/web/internal/item/selector/AssetEntryItemSelectorView.java
@@ -15,7 +15,6 @@
 package com.liferay.asset.browser.web.internal.item.selector;
 
 import com.liferay.asset.browser.web.internal.display.context.AssetBrowserDisplayContext;
-import com.liferay.asset.browser.web.internal.display.context.AssetBrowserManagementToolbarDisplayContext;
 import com.liferay.asset.kernel.service.AssetEntryLocalService;
 import com.liferay.asset.util.AssetHelper;
 import com.liferay.depot.service.DepotEntryService;
@@ -24,7 +23,6 @@ import com.liferay.item.selector.ItemSelectorView;
 import com.liferay.item.selector.ItemSelectorViewDescriptorRenderer;
 import com.liferay.item.selector.criteria.AssetEntryItemSelectorReturnType;
 import com.liferay.item.selector.criteria.asset.criterion.AssetEntryItemSelectorCriterion;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.Portal;
@@ -35,7 +33,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 
-import javax.portlet.PortletException;
 import javax.portlet.PortletURL;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
@@ -80,39 +77,29 @@ public class AssetEntryItemSelectorView
 			PortletURL portletURL, String itemSelectedEventName, boolean search)
 		throws IOException, ServletException {
 
-		try {
-			HttpServletRequest httpServletRequest =
-				(HttpServletRequest)servletRequest;
+		HttpServletRequest httpServletRequest =
+			(HttpServletRequest)servletRequest;
 
-			RenderRequest renderRequest =
-				(RenderRequest)httpServletRequest.getAttribute(
-					JavaConstants.JAVAX_PORTLET_REQUEST);
-			RenderResponse renderResponse =
-				(RenderResponse)httpServletRequest.getAttribute(
-					JavaConstants.JAVAX_PORTLET_RESPONSE);
+		RenderRequest renderRequest =
+			(RenderRequest)httpServletRequest.getAttribute(
+				JavaConstants.JAVAX_PORTLET_REQUEST);
+		RenderResponse renderResponse =
+			(RenderResponse)httpServletRequest.getAttribute(
+				JavaConstants.JAVAX_PORTLET_RESPONSE);
 
-			AssetBrowserDisplayContext assetBrowserDisplayContext =
-				new AssetBrowserDisplayContext(
-					_assetEntryLocalService, _assetHelper,
-					assetEntryItemSelectorCriterion, _depotEntryService,
-					httpServletRequest, _portal, portletURL, renderRequest,
-					renderResponse);
+		AssetBrowserDisplayContext assetBrowserDisplayContext =
+			new AssetBrowserDisplayContext(
+				_assetEntryLocalService, _assetHelper,
+				assetEntryItemSelectorCriterion, _depotEntryService,
+				httpServletRequest, _portal, portletURL, renderRequest,
+				renderResponse);
 
-			_itemSelectorViewDescriptorRenderer.renderHTML(
-				httpServletRequest, servletResponse,
-				assetEntryItemSelectorCriterion, portletURL,
-				itemSelectedEventName, search,
-				new AssetEntryItemSelectorViewDescriptor(
-					httpServletRequest, assetBrowserDisplayContext,
-					new AssetBrowserManagementToolbarDisplayContext(
-						httpServletRequest,
-						_portal.getLiferayPortletRequest(renderRequest),
-						_portal.getLiferayPortletResponse(renderResponse),
-						assetBrowserDisplayContext)));
-		}
-		catch (PortalException | PortletException exception) {
-			throw new ServletException(exception);
-		}
+		_itemSelectorViewDescriptorRenderer.renderHTML(
+			httpServletRequest, servletResponse,
+			assetEntryItemSelectorCriterion, portletURL, itemSelectedEventName,
+			search,
+			new AssetEntryItemSelectorViewDescriptor(
+				httpServletRequest, assetBrowserDisplayContext, portletURL));
 	}
 
 	private static final List<ItemSelectorReturnType>
