+++ modules/apps/asset/asset-browser-web/src/main/java/com/liferay/asset/browser/web/internal/item/selector/AssetEntryItemSelectorViewDescriptor.java
@@ -15,16 +15,25 @@
 package com.liferay.asset.browser.web.internal.item.selector;
 
 import com.liferay.asset.browser.web.internal.display.context.AssetBrowserDisplayContext;
-import com.liferay.asset.browser.web.internal.display.context.AssetBrowserManagementToolbarDisplayContext;
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.ItemSelectorViewDescriptor;
 import com.liferay.item.selector.criteria.AssetEntryItemSelectorReturnType;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.HtmlUtil;
+import com.liferay.portal.kernel.util.WebKeys;
 
 import java.util.List;
+
+import javax.portlet.PortletURL;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -37,13 +46,14 @@ public class AssetEntryItemSelectorViewDescriptor
 	public AssetEntryItemSelectorViewDescriptor(
 		HttpServletRequest httpServletRequest,
 		AssetBrowserDisplayContext assetBrowserDisplayContext,
-		AssetBrowserManagementToolbarDisplayContext
-			assetBrowserManagementToolbarDisplayContext) {
+		PortletURL portletURL) {
 
 		_httpServletRequest = httpServletRequest;
 		_assetBrowserDisplayContext = assetBrowserDisplayContext;
-		_assetBrowserManagementToolbarDisplayContext =
-			assetBrowserManagementToolbarDisplayContext;
+		_portletURL = portletURL;
+
+		_themeDisplay = (ThemeDisplay)httpServletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
 	}
 
 	@Override
@@ -53,8 +63,64 @@ public class AssetEntryItemSelectorViewDescriptor
 
 	@Override
 	public List<DropdownItem> getFilterNavigationDropdownItems() {
-		return _assetBrowserManagementToolbarDisplayContext.
-			getFilterNavigationDropdownItems();
+		long[] groupIds = _assetBrowserDisplayContext.getSelectedGroupIds();
+
+		if (groupIds.length <= 1) {
+			return DropdownItemListBuilder.add(
+				dropdownItem -> {
+					dropdownItem.setActive(
+						_assetBrowserDisplayContext.isSearchEverywhere());
+					dropdownItem.setHref(_portletURL, "scope", "everywhere");
+					dropdownItem.setLabel(
+						LanguageUtil.get(_httpServletRequest, "everywhere"));
+				}
+			).add(
+				dropdownItem -> {
+					dropdownItem.setActive(
+						!_assetBrowserDisplayContext.isSearchEverywhere());
+					dropdownItem.setHref(_portletURL, "scope", "current");
+					dropdownItem.setLabel(_getCurrentScopeLabel());
+				}
+			).build();
+		}
+
+		return new DropdownItemList() {
+			{
+				add(
+					dropdownItem -> {
+						dropdownItem.setActive(
+							_assetBrowserDisplayContext.isSearchEverywhere());
+						dropdownItem.setHref(
+							_portletURL, "scope", "everywhere", "groupId",
+							null);
+						dropdownItem.setLabel(
+							LanguageUtil.get(
+								_httpServletRequest, "everywhere"));
+					});
+
+				for (long groupId : groupIds) {
+					Group group = GroupLocalServiceUtil.fetchGroup(groupId);
+
+					if (group == null) {
+						continue;
+					}
+
+					add(
+						dropdownItem -> {
+							dropdownItem.setActive(
+								_assetBrowserDisplayContext.getGroupId() ==
+									groupId);
+							dropdownItem.setHref(
+								_portletURL, "groupId", groupId, "scope",
+								"current");
+							dropdownItem.setLabel(
+								HtmlUtil.escape(
+									group.getDescriptiveName(
+										_themeDisplay.getLocale())));
+						});
+				}
+			}
+		};
 	}
 
 	@Override
@@ -85,9 +151,29 @@ public class AssetEntryItemSelectorViewDescriptor
 		return true;
 	}
 
+	private String _getCurrentScopeLabel() {
+		Group group = _themeDisplay.getScopeGroup();
+
+		if (group.isSite()) {
+			return LanguageUtil.get(_httpServletRequest, "current-site");
+		}
+
+		if (group.isOrganization()) {
+			return LanguageUtil.get(
+				_httpServletRequest, "current-organization");
+		}
+
+		if (group.isDepot()) {
+			return LanguageUtil.get(
+				_httpServletRequest, "current-asset-library");
+		}
+
+		return LanguageUtil.get(_httpServletRequest, "current-scope");
+	}
+
 	private final AssetBrowserDisplayContext _assetBrowserDisplayContext;
-	private final AssetBrowserManagementToolbarDisplayContext
-		_assetBrowserManagementToolbarDisplayContext;
 	private final HttpServletRequest _httpServletRequest;
+	private final PortletURL _portletURL;
+	private final ThemeDisplay _themeDisplay;
 
 }
